==> 06 Gene level analysis Table 1.R <==
#generates the list of the 25 most significant up and down regulated genes for each tumor type.
#generates composite figures showing expression of these genes.
#generates lists of genes that could be used as single or paired markers for tumor types.
#generates information regarding the rlog expression and l2fc for these genes

library(stringr)
library(edgeR)
library(DESeq2)
library(ggplot2)
library(ggrepel)
library(FactoMineR)
library(factoextra)
library(gplots)
library(plotly)
library(ggpattern)
library(gridExtra)

#most of this was originally in the 02 Tumor type IDs.R script, but has been put here for clarity
plotGene<-function(gene, values, groups, types, box = TRUE, points = TRUE, ptcolor = NA, IDs = "", doprint = TRUE, doreturn = FALSE, addLegend = T, hideAxisTitles = NA, ptSize = 5){
  title<-gene
  title<-str_remove(title, "^gene-")
  mpDat<-data.frame(value = as.numeric(values),
                    group = as.factor(groups),
                    type = as.factor(types))
  colnames(mpDat)<-c("value","group","type")
  
  p<-ggplot(mpDat, aes(y = value, x = group, fill = group))
  if (box) p <- p +
    geom_boxplot_pattern(lwd = 1, 
                         color = "black", pattern_fill = "black", 
                         pattern_density = .1, outlier.shape = NA, aes(pattern = group), show.legend = addLegend) +
    scale_pattern_manual(values = c(FS = "none", PNST = "stripe", PWT = "none")) +
    scale_fill_manual(values = c(FS = "white",PNST = "white",PWT = "grey"))
  p
  
  if (points){
    if (is.na(ptcolor))
      p <- p+geom_jitter(shape=16, position=position_jitter(width = 0.2, seed = 1), size = ptSize, show.legend = F)
    else if (ptcolor == "types")
      p<- p+geom_jitter(shape=16, position=position_jitter(width = 0.2, seed = 1), size = ptSize, aes(color = type), show.legend = F)
    else
      p <- p+geom_jitter(shape=16, position=position_jitter(width = 0.2, seed = 1), size = ptSize, color = ptcolor, show.legend = F)
  }
  
  if (IDs[1] != ""){
    p <- p + geom_text(aes(label = IDs), position = position_jitter(width = 0.2, seed = 1), color = "red")
  }
  
  p<-p +
    theme_bw()+
    theme(axis.text.x = element_text(angle = 45, hjust = 1)) +
    ggtitle(title)+
    theme (
      plot.title = element_text(size=16,face="bold", hjust = 0.5),
      axis.text.y = element_text(size=16,face="bold"),
      axis.title.y = element_text(size=16,face="bold"),
      axis.text.x = element_text(size=16,face="bold"),
      axis.title.x = element_text(size=16,face="bold"),
      axis.line = element_line(size = 1.5),
      axis.ticks = element_line(size = 1.5),
      legend.text = element_text(face = "bold", size = 16),
      legend.title = element_text(size=16, face="bold")
    ) +
    theme(plot.margin=unit(c(1,1,1,1),"cm"))
  
  if (hideAxisTitles == "X"){
    p<-p + theme(axis.title.x = element_blank())
  } else
    if(hideAxisTitles == "Y"){
      p<-p + theme(axis.title.y = element_blank())
    } else
      if(hideAxisTitles == "both"){
        p<-p + theme(axis.title.y = element_blank(),
                     axis.title.x = element_blank())
      }
  
  if (doprint){
    print(p)
  }
  if(doreturn){
    return(p)
  }
}


MakePNGPlotTable<-function(geneNames,myFileName){
  myPlots<-list()
  
  for ( i in 1:length(geneNames)){
    gene<-paste0("gene-",geneNames[i])
    #print(gene)
    p<-plotGene(gene,mCPM[gene,], metaD$Histology.group, metaD$Histology.group, addLegend = F, hideAxisTitles = "both", doreturn = T, doprint = F, ptSize = 3)
    #print(p)
    myPlots[[i]]<-p
  }
  png(filename = myFileName, width = 1200, height = 1200)
  do.call("grid.arrange", c(myPlots[1:25], ncol = 5))
  dev.off()
}

####################################################################################################
load("C:\\Users\\idaho\\Dropbox\\Shelden lab\\Canine SARC analysis\\Input Data\\all_data.rda")
load("C:\\Users\\Eric Shelden\\Dropbox\\Shelden lab\\Canine SARC analysis\\Input Data\\all_data.rda")

mDat<-all_data
colnames(mDat)
mDat<-mDat[,1:16] #analyze just the cSTS
colnames(mDat)

#sigGenes.rda contains FS.genes.sig, PNST.genes.sig, PWT.genes.sig, mCPM, generated by "02 Tumor type IDs.R"
setwd("C:\\Users\\idaho\\Dropbox\\Shelden lab\\Canine SARC analysis\\Results\\Tumor type specific gene expression\\Eric")
setwd("C:\\Users\\Eric Shelden\\Dropbox\\Shelden lab\\Canine SARC analysis\\Results\\Tumor type specific gene expression\\Eric")
load("sigGenes.rda")

metaD<-read.csv("C:\\Users\\idaho\\Dropbox\\Shelden lab\\Canine SARC analysis\\Input Data\\Subject Data\\STS Data for study-ES-reassigned.csv")
metaD<-read.csv("C:\\Users\\Eric Shelden\\Dropbox\\Shelden lab\\Canine SARC analysis\\Input Data\\Subject Data\\STS Data for study-ES-reassigned.csv")
dim(metaD) #16 x 43
rownames(metaD)<-metaD$Sample.ID
#reorder the metaD data
metaD<-metaD[colnames(mCPM),]
all.equal(colnames(mCPM),metaD$Sample.ID) #TRUE
table(metaD$Histology.group) #4 FS, 6 PNST, 6 PWT

FS.samples <- rownames(metaD[metaD$Histology.group == "FS",])
PNST.samples <- rownames(metaD[metaD$Histology.group == "PNST",])
PWT.samples <- rownames(metaD[metaD$Histology.group == "PWT",])

########### 3-3-22 Top 25 up and down genes for each tumor type. ###################################

dim(FS.genes.sig) #2090
dim(PWT.genes.sig) #875
dim(PNST.genes.sig) #291

#find the top 25 most significant up and down regulated DEGs for each tumor type
FS.genes.sig<-FS.genes.sig[order(FS.genes.sig$padj, decreasing = F),]
annotated.FS.genes.sig<-FS.genes.sig[!grepl("^gene-LOC",rownames(FS.genes.sig)),]
sum(annotated.FS.genes.sig$log2FoldChange > 0) #1138
sum(annotated.FS.genes.sig$log2FoldChange < 0) #499
top.sig.FS.up.genes<-annotated.FS.genes.sig[annotated.FS.genes.sig$log2FoldChange > 0,]
top.sig.FS.up.genes<-top.sig.FS.up.genes[1:25,]
top.sig.FS.up.genes
top.sig.FS.dn.genes<-annotated.FS.genes.sig[annotated.FS.genes.sig$log2FoldChange < 0,]
top.sig.FS.dn.genes<-top.sig.FS.dn.genes[1:25,]
top.sig.FS.dn.genes

PNST.genes.sig<-PNST.genes.sig[order(PNST.genes.sig$padj, decreasing = F),]
annotated.PNST.genes.sig<-PNST.genes.sig[!grepl("^gene-LOC",rownames(PNST.genes.sig)),]
sum(annotated.PNST.genes.sig$log2FoldChange > 0) #78
sum(annotated.PNST.genes.sig$log2FoldChange < 0) #132
top.sig.PNST.up.genes<-annotated.PNST.genes.sig[annotated.PNST.genes.sig$log2FoldChange > 0,]
top.sig.PNST.up.genes<-top.sig.PNST.up.genes[1:25,]
top.sig.PNST.up.genes
top.sig.PNST.dn.genes<-annotated.PNST.genes.sig[annotated.PNST.genes.sig$log2FoldChange < 0,]
top.sig.PNST.dn.genes<-top.sig.PNST.dn.genes[1:25,]
top.sig.PNST.dn.genes

PWT.genes.sig<-PWT.genes.sig[order(PWT.genes.sig$padj, decreasing = F),]
annotated.PWT.genes.sig<-PWT.genes.sig[!grepl("^gene-LOC",rownames(PWT.genes.sig)),]
sum(annotated.PWT.genes.sig$log2FoldChange > 0) #120
sum(annotated.PWT.genes.sig$log2FoldChange < 0) #420
top.sig.PWT.up.genes<-annotated.PWT.genes.sig[annotated.PWT.genes.sig$log2FoldChange > 0,]
top.sig.PWT.up.genes<-top.sig.PWT.up.genes[1:25,]
top.sig.PWT.up.genes
top.sig.PWT.dn.genes<-annotated.PWT.genes.sig[annotated.PWT.genes.sig$log2FoldChange < 0,]
top.sig.PWT.dn.genes<-top.sig.PWT.dn.genes[1:25,]
top.sig.PWT.dn.genes

#Generate plots for each set of 25 genes. Uses function MakePNGPlotTable, defined above.
geneNames<-rownames(top.sig.FS.up.genes) %>% str_remove("^gene-")
MakePNGPlotTable(geneNames, "FSUP.png")

geneNames<-rownames(top.sig.FS.dn.genes) %>% str_remove("^gene-")
MakePNGPlotTable(geneNames, "FSDN.png")

geneNames<-rownames(top.sig.PNST.up.genes) %>% str_remove("^gene-")
MakePNGPlotTable(geneNames, "PNSTUP.png")

geneNames<-rownames(top.sig.PNST.dn.genes) %>% str_remove("^gene-")
MakePNGPlotTable(geneNames, "PNSTDN.png")

geneNames<-rownames(top.sig.PWT.up.genes) %>% str_remove("^gene-")
MakePNGPlotTable(geneNames, "PWTUP.png")

geneNames<-rownames(top.sig.PWT.dn.genes) %>% str_remove("^gene-")
MakePNGPlotTable(geneNames, "PWTDN.png")


################## Analyze individual genes for potential markers ##########################
FS.samples <- rownames(metaD[metaD$Histology.group == "FS",])
PNST.samples <- rownames(metaD[metaD$Histology.group == "PNST",])
PWT.samples <- rownames(metaD[metaD$Histology.group == "PWT",])
thresh<-log2(1.8)

#marker found if the lowest expression value in FS.smaples is greater than the largest value
#of all other samples plus the minimum (thresh) fold difference we set
thresh<-log2(2)
for(i in 1:nrow(mCPM)){
  if (min(mCPM[i,FS.samples]) > (1+max(mCPM[i,c(PNST.samples,PWT.samples)]))){
    print(rownames(mCPM)[i])
  }
}

for(i in 1:nrow(mCPM)){
  if (min(mCPM[i,PNST.samples]) > (thresh+max(mCPM[i,c(FS.samples,PWT.samples)]))){
    print(rownames(mCPM)[i])
  }
}

for(i in 1:nrow(mCPM)){
  if (min(mCPM[i,PWT.samples]) > (thresh+max(mCPM[i,c(FS.samples,PNST.samples)]))){
    print(rownames(mCPM)[i])
  }
}

#downregulated?
for(i in 1:nrow(mCPM)){
  if (max(mCPM[i,PNST.samples])+thresh < (min(mCPM[i,c(FS.samples,PWT.samples)]))){
    print(rownames(mCPM)[i])
  }
}

for(i in 1:nrow(mCPM)){
  if (max(mCPM[i,PWT.samples])+thresh < (min(mCPM[i,c(FS.samples,PNST.samples)]))){
    print(rownames(mCPM)[i])
  }
}

#Analyze pairs of genes as markers
#basically, the stuff above asks if all samples in a group are greater than all samples in another group using some cutoff.
#below, we will instead ask if all samples in a group are above some cutoff in one gene OR a second gene.
thresh<-log2(1.75)

for (ID1 in 1:nrow(PNST.genes.sig)){
  for (ID2 in ID1:nrow(PNST.genes.sig)){
    gene1<-rownames(PNST.genes.sig)[ID1]
    gene2<-rownames(PNST.genes.sig)[ID2]
    #print(paste(gene1,gene2))
    found<-T
    #test for sensitivity
    for (sample in PNST.samples){
      #both genes in a sample have to be greater than threshold over expression of all other samples
      #if it's less for either, the gene does not pass sensitivity testing
      if (mCPM[gene1,sample] < thresh+max(mCPM[gene1,c(FS.samples,PWT.samples)]) &
          mCPM[gene2,sample] < thresh+max(mCPM[gene2,c(FS.samples,PWT.samples)])){
        found<-F
      }
    } #checked all samples for sensitivity
    
    if (found){
      print(paste(gene1,gene2))
    }
  }
}

for (ID1 in 1:nrow(PWT.genes.sig)){
  for (ID2 in ID1:nrow(PWT.genes.sig)){
    gene1<-rownames(PWT.genes.sig)[ID1]
    gene2<-rownames(PWT.genes.sig)[ID2]
    #print(paste(gene1,gene2))
    found<-T
    #test for sensitivity
    for (sample in PWT.samples){
      #both genes in a sample have to be greater than threshold over expression of all other samples
      #if it's less for either, the gene does not pass sensitivity testing
      if (mCPM[gene1,sample] < thresh+max(mCPM[gene1,c(FS.samples,PNST.samples)]) &
          mCPM[gene2,sample] < thresh+max(mCPM[gene2,c(FS.samples,PNST.samples)])){
        found<-F
      }
    } #checked all samples for sensitivity
    
    if (found){
      print(paste(gene1,gene2))
    }
  }
}

############################### Statistics on expression of top genes #####################
#reviewer 1 wants more information on expression of top 25 up and down genes for each tumor.
#specifically, p value, fold change and expression level.
#plot fold change vs expression level?
mNames<-rownames(top.sig.FS.up.genes)

mean(mCPM["gene-SNAI1",FS.samples]) #3.64
mean(mCPM["gene-SNAI1",PNST.samples]) #3.64
mean(mCPM["gene-SNAI1",PWT.samples]) #3.64
mean(mCPM["gene-SNAI1",]) #3.50

mean(mCPM["gene-KRT13",]) #-1.47

allSamples<-c(FS.samples,PNST.samples,PWT.samples)

myData<-top.sig.FS.up.genes
mySamples<-FS.samples
DoAnalyze(myData, mySamples)
#OR
myData<-top.sig.PNST.up.genes
mySamples<-PNST.samples
DoAnalyze(myData, mySamples)
#OR
myData<-top.sig.PWT.up.genes
mySamples<-PWT.samples
DoAnalyze(myData, mySamples)
#OR for down regulated genes
myData<-top.sig.FS.dn.genes
mySamples<-FS.samples
DoAnalyze(myData, mySamples)
#OR
myData<-top.sig.PNST.dn.genes
mySamples<-PNST.samples
DoAnalyze(myData, mySamples)
#OR
myData<-top.sig.PWT.dn.genes
mySamples<-PWT.samples
DoAnalyze(myData, mySamples)

DoAnalyze<-function(myData, mySamples){
  otherSamples<-setdiff(allSamples,mySamples)
  mNames<-rownames(myData)
  mTable<-as.data.frame(matrix(0,nrow = 25, ncol = 3))
  rownames(mTable)<-mNames
  colnames(mTable)<-c("expression", "l2fc", "other-expression")
  
  for (i in 1:25){
    mTable[i,1]<-mean(mCPM[rownames(mTable)[i],mySamples])
    mTable[i,2]<-myData[i,]$log2FoldChange
    mTable[i,3]<-mean(mCPM[rownames(mTable)[i],otherSamples])
  }
  
  #x is expression, y is log2foldchange
  print("order by l2fc")
  print(mTable[order(mTable$l2fc),])
  print("")
  print("order by expression")
  print(mTable[order(mTable$expression),])
  print("")
  print("lowest tumor specific l2fc")
  print(mTable[which(mTable$l2fc == min(mTable$l2fc)),])
  print("average tumor specific l2fc")
  print(mean(mTable$l2fc))
  print("lowest tumor specific rlog expression")
  print(mTable[which(mTable$expression == min(mTable$expression)),])
  print("average tumor specific rlog expression")
  print(mean(mTable$expression))
  print("average other rlog expression")
  print(mean(mTable$`other-expression`))
}
